Adds versioning to api
