+++ config/routes.rb
@@ -1,10 +1,12 @@
 Rails.application.routes.draw do
-  resources :characters do
-  end
-
-  resources :episodes do
-  end
-
-  resources :locations do
+  scope module: :v1, constraints: ApiVersion.new('v1', true) do
+    resources :characters do
+    end
+  
+    resources :episodes do
+    end
+  
+    resources :locations do
+    end
   end
 end
