+++ spec/controllers/locations_controller_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
-RSpec.describe LocationsController, type: :controller do
+RSpec.describe V1::LocationsController, type: :controller do
 
 end
